+++ lib/src/core/better_player_with_controls.dart
@@ -151,7 +151,8 @@ class _BetterPlayerWithControlsState extends State<BetterPlayerWithControls> {
   }
 
   Widget _buildPlaceholder(BetterPlayerController betterPlayerController) {
-    return betterPlayerController.betterPlayerConfiguration.placeholder ??
+    return betterPlayerController.betterPlayerDataSource.placeholder ??
+        betterPlayerController.betterPlayerConfiguration.placeholder ??
         Container();
   }
 
@@ -283,6 +284,11 @@ class _BetterPlayerVideoFitWidgetState
                 widget.betterPlayerController.hasCurrentDataSourceStarted;
           });
         }
+      }
+      if (event.betterPlayerEventType == BetterPlayerEventType.setupDataSource){
+        setState(() {
+          _started = false;
+        });
       }
     });
   }